requirements.txt
numpy>=1.21.0
torch>=2.0.0
webrtcvad>=2.0.10
pyaudio>=0.2.13
sounddevice>=0.4.6
faster-whisper>=0.9.0
TTS>=0.22.0
llama-cpp-python>=0.2.11
python-dotenv>=1.0.0
streamlit>=1.32.0
anthropic>=0.18.1
openai>=1.12.0
requests>=2.31.0 